internal/ui: refactor settings stuff into its own file
