+++ internal/ui/app.go
@@ -45,28 +45,6 @@ type App struct {
 	peerPages  map[key.NodePublic]*peerPage
 }
 
-func (a *App) showPreferences() {
-	if a.settings == nil {
-		a.toast("Settings schema not found")
-		return
-	}
-
-	win := NewPreferencesWindow()
-	a.settings.Bind("tray-icon", win.UseTrayIcon.Object, "active", gio.SettingsBindDefault)
-	a.settings.Bind("polling-interval", win.PollingIntervalAdjustment.Object, "value", gio.SettingsBindDefault)
-	a.settings.Bind("control-plane-server", win.ControlServer.Object, "text", gio.SettingsBindGet)
-	win.SetTransientFor(&a.win.Window)
-	win.ConnectCloseRequest(func() bool {
-		// Anything that shouldn't update every time the value of the
-		// widget changes should cause an update here manually instead.
-		a.settings.SetString("control-plane-server", win.ControlServer.Text())
-		return false
-	})
-	win.Show()
-
-	a.app.AddWindow(&win.Window.Window)
-}
-
 // showAbout shows the app's about dialog.
 func (a *App) showAbout() {
 	dialog := adw.NewAboutWindow()
@@ -211,49 +189,6 @@ func (a *App) init(ctx context.Context) {
 	})
 
 	a.initSettings(ctx)
-}
-
-func (a *App) initSettings(ctx context.Context) {
-	nonreloc, reloc := gio.SettingsSchemaSourceGetDefault().ListSchemas(true)
-	if !slices.Contains(nonreloc, appID) && !slices.Contains(reloc, appID) {
-		goto init
-	}
-
-	a.settings = gio.NewSettings(appID)
-	a.settings.ConnectChanged(func(key string) {
-		switch key {
-		case "tray-icon":
-			if a.settings.Boolean("tray-icon") {
-				go tray.Start(func() { a.initTray(ctx) })
-				return
-			}
-			tray.Stop()
-
-		case "polling-interval":
-			a.poller.SetInterval() <- a.getInterval()
-
-		case "control-plane-server":
-			url := a.settings.String("control-plane-server")
-			err := a.TS.SetControlURL(ctx, url)
-			if err != nil {
-				slog.Error("update control plane server URL", "err", err, "url", url)
-				return
-			}
-			a.poller.Poll() <- struct{}{}
-		}
-	})
-
-init:
-	if (a.settings == nil) || a.settings.Boolean("tray-icon") {
-		go tray.Start(func() { a.initTray(ctx) })
-	}
-}
-
-func (a *App) getInterval() time.Duration {
-	if a.settings == nil {
-		return 5 * time.Second
-	}
-	return time.Duration(a.settings.Double("polling-interval") * float64(time.Second))
 }
 
 func (a *App) startTS(ctx context.Context) error {
